Adding basic logging support.
This implement a logging plugin that log to a file in the log directory. It is
assumed that the user use something like logrotate(8) to rotate the logs. In
addition, the program name is added to the information given to the plugin
since that is used to derive the log file name.

This patch also adds support for plugins to provide interface files that are
installed together with the plugin.