WL8480 Add mysqlrouter application
We add the main application of MySQL Router to the
repository. It is named `mysqlrouter` and its source is
located under `src/router`.

Harness is required to be able to build MySQL Router. You
can use the CMake variable `WITH_HARNESS` to point to the
source of Harness or make sure it is available in the
root repository of MySQL Router (possibly as symbolic
link) as `harness/`.

To build the executable:

    shell> mkdir build
    shell> cd build
    shell> cmake ..
    shell> make

To install the application:

    shell> make install

To install in a different location:

    shell> cmake .. -DCMAKE_INSTALL_PREFIX=/opt/mysqlrouter
    shell> make
    shell> make install

To build the documentation, after which it is available in the
build tree as `doc/html/index.html`:

    shell> make doc

To build and run tests, make sure that you have Google C++ Testing
Framework available. You can specify the location of the testing
framework using the CMake variable GTEST_ROOT or make sure it is
available in the root repository of MySQL Router (possibly as
symbolic link) as `gtest/`.

    shell> cd build
    shell> cmake .. -DENABLE_TESTS -DGTEST_ROOT=/path/to/gtest
    shell> make
    shell> make test

To run the mysqlrouter application you need to point to a valid
configuration or make sure a configuration file is available
in a default location. To check for these locations and get, use
the `--help` (or `-h`) command line argument:

    shell> mysqlrouter --help

It is possible to retrieve the version number of using the the
command line argument `--version` (or `-v`):

    shell> mysqlrouter --version

To point to a specific configuration file, use the the `--config`
option or `-c`:

    shell> mysqlrouter --config /path/to/mysqlrouter.ini

When the argument `--config` is used, no default file locations will
be used.

It is possible to ammend or read extra configuration files with the
command line argument `--extra-config` or `-a`. Reading extra
configuration files needs at least a default or a configuration file
given using the `--config` option.

It is also possible to specify multiple times the `--config` and/or
the `--extra-config` option. For example:

    shell> mysqlrouter -c c1.cnf -a a1.cnf -c c2.cnf -a a2.cnf

In the above command line, configuration file c1.cnf and c2.cnf will
be read first. Then the extra files a1.cnf and a2.cnf are read. Files are
read according to the order they are specified on the command line.