Fixing memory allocation issue in configuration parser.
When using ConfigSection inside the configuration parser, there were a memory
allocation issue resulting from a copy-construction of the instance.

This patch eliminate the problem by deleting the copy-constructor and
copy-assignment operator and ensure that the ConfigSection is constructed
in-place.

When requesting a section by name there could be several sections with the same
name but different keys.

This patch fixes the issue by having the function searching by name return a
sequence of sections rather than a single section.