Adding support for reading multiple configuration file entries.
This patch add functionality to read configurations from either a file, or from
a directory, commonly referred to as a *configuration entry*. If reading from a
directory, all configuration files in the directory will be merged together to
form a configuration entry. Within each configuration entry, options are
assumed to not conflict and if values are given to options, an error is thrown.
However, sequences of entries can be read and is then assumed to overwrite the
existing configuration options.

The patch also split out Loader::read() from constructor. This means that
constructor no longer accept configuration file for reading and that the read()
function should be used after construction.