Change plugin callback signature to pass configuration section.
This patch change the functions init() and start() so that init() is called
once for each plugin and one thread is spawned for each configuration section.

For the init() call, a pointer to the application information is passed, and
for the start() function, a pointer to the configuration section is passed.

For example, given this configuration file:

    [router:first]
    source = ...
    destination = ...

    [router:second]
    source = ...
    destionation = ...

The function init() will be called once and two threads will be spawned using
start(): one for `router:first` and one for `router:second`.

In addition, the patch remove the reference wrappers used in the section lists
since that caused more problems than necessary and added a default for the 'library' option.

The section lists passed back from functions in Config now use
pointers to the sections in the configuration.

The default value of the 'library' option is the section name (without the
section key) if the option is not given.