BUG21629154 Fix compile/tests issues on Linux
We fix the configuration files for testing using the new default
options such as logging_folder instead of logdir.

We remove most of the compiler flags when testing is enabled to allow
GMock/GTest to compile.

The tests checking license and copyright are improved to show which
files in correct and why. The scripts will now show a list of failures.