BUG21641430 Add Debian packaging support
We add support for creating Debian packages for the following
Linux distributions:

    * Debian 8 (jessie)
    * Ubuntu 15.04 (vivid)
    * Ubuntu 14.04 (trusty)

The Debian packaging files can be found per distribution in the
packaging folder. The build_deb.sh script was added to make it
easier to create packages. Provided dependencies (devscripts,
debhelper and realpath), the script can be used like this:

 $ mkdir build && cd build
 $ cmake .. && make package_source
 $ sh ../packaging/debian/build_deb.sh mysql-router-2.0.0.tar.gz

The resulting Debian packages should be available in the current
working directory.

The installed init script is /etc/init.d/mysqlrouter.

Use chuid in init script; add postinst