BUG21618845 Fix handling unreadable configuration files
We now exit the application with an error when one of configuration
files given through command line options is not usable.

We also remove the possibility to use the --config option more than
once. An error is now reported. To read extra information, the
--extra-config option should be used.

Unit tests were added and, where needed, updated.