Introducing GTest for unit tests.
Switching to using GTest for the unit tests rather than the home-grown system.

It also fixes two bugs that were discovered in the transition.

If a check for a non-existant section was done using find_range_first it
checked for one-past-end of container rather than that the range returned was
empty.

This patch fixes the issue by checking if the range is empty.

The Loader class did not define a virtual destructor, but was inherited.