Adding FindGMock.cmake to repository
This will search for the GMock library in the standard path and if not found,
report an error.

If the GMock library is installed in a non-standard location, the variable
GMOCK_ROOT can be used either as an environment variable or as a CMake
variable to refer to the installation root of GMock.