Adding test flag for enabling tests.
The flag ENABLE_TESTS is added (default is OFF) that allow you to enable unit
tests to be executed.

In addition, the patch move all tests to tests/ and hard-coding path to test
files, including plugins.

Before this patch, solutions setting the HARNESS_NAME would have test files
installed in different locations. To avoid conflicting with solutions using the
harness, the path to the plugin directories and other test directories are now
hard-coded.

To handle this, the ability to provide a destination prefix for plugin
installation paths where added to add_plugin. This is used by the tests to
hard-code the name of the destination prefix so that there is no conflict with
real plugins used by the solution.