Using XSI-conforming functions and taking care of return value from strerror_r.
This patch set _XOPEN_SOURCE to indicate that we are POSIX.1-2001 compliant and
XSI-conforming. That will automatically set _POSIX_C_SOURCE to the correct
value and ensure that we use the XSI version of strerror_r.

This pre-processor symbol is set in the main CMakeLists.txt file when the
compiler is either GNU or CLang.

Not handling the return value from strerror_r generate a warning, and since we
have errors-on-warnings, this block the Release build. This patch check the
return value to ensure that strerror_r do not fail (!) and print a generic
error in the case that it does fail.