WL8479 Connection Routing (1)
We add the Connection Routing module and plugin, enabling routing
MySQL traffic from clients to an appropriated MySQL server.

Connection routing does not parse any packets and we require each
route to be running on a different TCP port. Connectors and client
will have to connect to the appropriated TCP port.

This first commit supports destinations to be provided as a comma
separated list of MySQL Servers.

For example, we want Router to listen on port 7001 and let clients
connected to the a MySQL Server configured as Master. When using mode
"read-write", routing will always try the first server before going
to the next, providing basic fail-over.

  [routing:A]
  bind_address = 0.0.0.0:7001
  destination = 127.0.0.1:3306,127.0.0.1:3307
  mode = READ-WRITE

When read-only mode is used, routing will go in a round-robin fashion
over all provided MySQL Servers. Each server is tried and not
excluded from the list.

  [routing:B]
  bind_address = 0.0.0.0:7002
  destination = 127.0.0.1:3306,slave1.example.com
  mode = read-only

Two timeouts can be provided for each route: wait_timeout and
connection_timeout.

* wait_timeout works like the MySQL Server's wait_timeout and will
disconnect clients which are idling for the given amount of seconds.
Default is 300.

* connect_timeout specifies how long we wait for the destination
MySQL Server to connect. When this timeout is reached, the next
server in the list will be tried. It is important to keep this
timeout low and the default is there set to 1 second.

Is is possible to limit the number of active connections which go
through a particular routing using the max_connections option. By
default it is set to 512.