Fixing issue with logger not flushing output.
There were a missing flush after a log message was written, causing nothing to
show up in the log file.

In addition, this patch remove the usage message on runtime errors (but still
keep it on incorrect invocation) for the 'harness' program and and move and
change the example configuration file to work with the keepalive plugin and use
the .ini suffix instead of the .cfg prefix.