Re-factoring the build system.
Dependencies between plugins, programs, and harness library was implicit and
handled by copying header files to the build directory and creating fake
targets for handling the plugin interface files. This caused subtle problems
when dependencies between plugins were not declared correctly, causing them to
succeed in some build and fail on others depending on what order the copy was
done.

This patch use the existing dependency system for CMake and declare include
files for plugins, libraries, and programs using INTERFACE_INCLUDE_DIRECTORIES
target property. This means that when a plugin is dependent on another plugin,
the include path will be propagated through the dependency chain and incorrect
dependency declarations will be discovered immediately.

In addition, the Harness test files where moved into the Harness directory so
that we have a uniform structure for all modules and plugins, and the
`add_plugin` macro was re-named to `add_harness_plugin` since it is now
dedicated to building Harness plugins. Frameworks based on the Harness then
need to define their own function to add any extra dependencies necessary.

The option ENABLE_HARNESS_PROGRAM was added, allowing the 'harness' program to
be excluded from the build and install, if it is not needed in the framework
using it.

All plugins are now written to the `lib/` directory under the binary output
directory. This includes the test plugins, which are also written to the same
directory. Note that the test plugins are *not* installed even if they are
built.

The patch also fixes some memory barrier issues and a missing return.