Changing default destination suffix for plugins to the Harness name.
The default destination suffix to use for Harness plugins was "harness", but
since we do not have a plugin path, we need to install all plugins in the same
directory. Hence the application using the framework decide on what destination
suffix to use by setting the variable `HARNESS_NAME` in the CMake build files.