Fix for assertion firing in logger plugin.
When running the logger plugin test, an assert was fired because `g_log_file`,
representing the log file, was null.

This was caused because the `keepalive.so` plugin kept a reference to the
`logger.so` located in the harness plugin directory but when the test ran, the
logger located in the test directory was used. This meant that two instances of
the `logger.so` plugin was present and initializing the logger just set the
`g_log_file` in the loaded one, while `keepalive.so` used the one it refrenced,
which was not initialized.

This patch fixes the problem by having all configuration files for the tests
refer to the harness plugin directory under the build root, which is shared by
all tests, instead of using a local version. This means that the same plugin
was loaded both through the reference from the `keepalive.so` plugin and the
dynamic loader.