Removing use of `SYSTEM` from `target_link_libraries`
The option `SYSTEM` was introduced in 2.8.12, but we have a requirement on
being able to build using 2.8.9 since we have platforms that do not support
later versions of CMake.

This patch remove the use of `SYSTEM` and use quoted includes for all
third-party include directives.

It also add a header for each section of includes to make the order clear.