Adding architecture descriptor to plugin structure.
This patch add an architecture descriptor to the plugin structure but do not
add any architecture checking. The reason is to ensure that plugins are written
the correct way so that adding architecture checks only require a re-compile of
the plugins.

The architecture descriptor is automatically generated from the platform and
compiler information.