Set the install rpath for plugins.
When plugins are being loaded, they need to be able to figure out the location
of other plugins. Since plugins are not installed in standard locations, this
means that the rpath need to be set for them when they are being installed.

This patch set the install rpath for plugins to be the install directory for
plugins, but rely on the standard lookup path to find the location of
libharness.so.