WL8479 Connection Routing (2)
We now are able to use the Fabric Cache in the configuration as
an alternative to the comma-separated-list of destination
servers.

Note: the 'destination' option was renamed to 'destinations'.

The following example shows a Fabric Cache plugin configuration
with key 'farm'. This can be used with in the URI when
configuring the Routing plugin:

  [fabric_cache:farm]
  address = your_fabric_node.example.com:32275
  user =
  password =

  [routing:homepage_writes_fabric]
  bind_address = 127.0.0.1:7001
  destinations = fabric+cache://farm/group/g1
  mode = read-write
  max_connections = 200

  [routing:homepage_reads_fabric]
  bind_address = 127.0.0.1:7002
  destinations = fabric+cache://farm/group/g1?allow_primary_reads=yes
  mode = read-only
  max_connections = 200
  wait_timeout = 100

A sample configuation is available in the docs folder.