BUG21771595 Fix config errors and exit instead of logging them
We exit the application when a configuration issues is noticed.
Instead of logging the error, the mysqlrouter application
terminates with a message starting with "Configuration error".

For example, when an invalid wait_timeout is given for the
Connection Routing plugin:

    $ mysqlrouter
    Configuration error: option wait_timeout in [routing] needs...

When a configuration error occurs, exit code 1 is returned.

Previously, errors shown by the mysqlrouter application where directed
STDOUT. We now use STDERR instead.

We add new test case for BUG21771595 and fix other tests.