BUG21899753 Make [logger] section optional
We make the [logger] section optional by adding it to the
configuration when it is missing. We also add the possibility
to define the log level. Default log level is INFO, which means
that debugging messages will not be logged.

Example:

  [logger]
  level = DEBUG

Valid values for the level option are FATAL, ERROR, WARNING, INFO,
or DEBUG (levels are case insensitive).