BUG21900022 Make [DEFAULT] section optional
We make the [DEFAULT] section in the configuration files optional.
The default values are defined at compile time and shown in the
output of the `mysqlrouter --help`.

The values for logging_folder, plugin_folder, runtime_folder and
config_folder will be different when using a different installation
layout (CMake -DINSTALL_LAYOUT option). When no installation layout
is specified (defaulting to STANDALONE), the defaults will be:

  $ /usr/local/bin/mysqlrouter --help
  ..
  Default configuration:
    config_folder  = /usr/local/etc/mysqlrouter
    logging_folder =
    plugin_folder  = /usr/local/lib/mysqlrouter
    runtime_folder = /usr/local/

When started, Router will now report on the console to which log file
it is writing messages.

The Debian-based packaging has been updated to reflect the new
defaults for the configuration and log location. We also simplified
the structure of `deb-*` folders so it is easier to maintain.