BUG21935219 Fix locations for the STANDALONE install layout
We rename the STANDALONE installation layout as DEFAULT and recreate
the STANDALONE with its original values.

The runtime and config folder are set to 'etc/' and 'run/'. The plugin
folder will be set to 'lib/mysqlrouter'. All three are relative to
the installation prefix.
The logging folder is kept empty so messages will go to the console.