Fix RPM and Debian packaging
We fix the RPM packaging installing libraries in the correct
location based on the architecture.

The spec file is now created using CMake `configure_file()` and will
make sure that the MySQL Router version stays up-to-date.

The Debian-based `rules` have been all updated to include the
substitution of the library folder.