Fix RPM and Debian packaging
We add a CMake variable which can be used when creating non-GPL
packages. It does need the internal repository which is a Git subtree
with prefix `internal`. Note that the README.txt and License.txt
files are changed in-source.

    $ cmake -DGPL=no

We cleaned up the Debian folders moving most of of the files from
per distribution to the common directory `deb-common`. This makes
maintaince easier.

We add a apparmor profile for Debian and Ubuntu.

We add the location ${CMAKE_BINARY_DIR}/../mysql-server when looking
for MySQL libraries and development files. This takes precendence to
the -DWITH_MYSQL location.

Both RPMs and Debian packages are now statically linking the MySQL
Client libraries using the CMake option -DWITH_STATIC=yes.

Dependency on libmysqlclient18/20 have been removed from the packages.

The RPM builds have been fix and improve so it will use the -DWITH_MYSQL
CMake option.