BUG#21944649: Setting RPATH correctly for standalone layouts.
When building using the STANDALONE layout, full paths were used. Since this is
used when building tarballs, the RPATH have to be set relative the shared
library and the executable.

This patch set the RPATH for shared libraries to `$ORIGIN/../lib` so that the
executable and shared libraries find the shared libraries in the tree, and the
RPATH for plugins to `$ORIGIN` so that plugins can find other plugins. (Note
that on OS X we use `@executable_path` instead of `$ORIGIN`).

In order to be able to reference locations relative to the location of the `mysqlrouter`
binary in the configuration file, the configuration variable `origin` was added
to the default section.  This is the directory part of `argv[0]`.

The patch also remove the output of defaults from `--help` since it was not
clear what the values represented.
(cherry picked from commit 5b2cbc0)