BUG21962350 Fix removing from quarantine
We fix the removal of servers from quarantine. We now correctly
remove servers from list and bail out when we reach the end of the list.

A test case was added for BUG21962350 as well as a flag making sure
that log is not being flooded with messages saying all servers are
quarantined.