BUG22057234 Default 127.0.0.1 for bind_address; add default_port
We add the possibility to define a TCP port through the option
default_port. We also set the default bind_address to '127.0.0.1'.
The following configuration is now the bare minimum to bind a
Connection Routing plugin to 127.0.0.1:7001:

  [routing]
  default_port = 7001

It is still possible to use the bind_address alone but then a port
is required.

  [routing]
  bind_address = 127.0.0.1:7001

New test cases were added and old ones updated.

Review comments (1)

Review comments (2); finish tests