BUG22024872 Password handling for Fabric Cache plugin
We are now prompting for the Fabric password when the Fabric Cache
plugin is used. An error is reported when the user has given a
password in the configuration file.

For testing purposes, it is possible to pipe the password to the
mysqlrouter application, thus not prompting for password. Note that
this only works when the Fabric Cache plugins use the same MySQL Fabric
instance.