BUG22045940 Fix connecting when Fabric started after Router
We remove the fetching of data in the FabricCache constructor. It
gave problems in situations when MySQL Fabric is not running and
Router is started before Fabric is available.

We make Router now correctly retry to connect and fetch data.

We also clean up some unneeded configuration options in the
Fabric Cache plugin.