BUG21972632 Fix getting next server when Fabric list is smaller
When the list of available managed servers in Fabric Cache is
is smaller and the current position was pointing to a managed
server that is not any more in the list, an uncatched exception
was thrown.
This problem could occure when when promoting and demoting a
HA Group in MySQL Fabric.

We add an additional check when the next-up index is still
available or not, after the list of managed servers has changed.