BUG22084430 Fix IPv6 addresses to use [] in logs
Problem

In some cases, the IPv6 address misses the square brackets around the IP
address. This is confusing when displayed with port.

Solution

In the constructor `TCPAddress(string, uint32_t)` the variable `ip_family_`
was being set to `Family::UNKNOWN` and so the condition
`if (ip_family_ == Family::IPV6)` in the function `TCPAddress::str()` was
failing resulting in IPV6 address being printed without square brackets.
Running the function `TCPAddress::detect_family()` in the constructor would
set `ip_family_` to an appropriate value.