BUG22074209 Add Router version to --help output
Problem

The output of the `--help` option does not show the version number of the
router. It should also show the platform distribution (if applicable) and
license type.

Solution

Including the call to `MySQLRouter::get_version_line()` in
`MySQLRouter::show_help()` would output the version information of
MySQL Router before the help text similar to other MySQL application
and projects.