WL#8957: Thread-safe Queue
This patch add an implementation of a thread-safe queue implemented using a
singly-linked list. To avoid contention, it has two mutexes: one for the head
of the queue and one for the tail of the queue. This allow threads adding
elements to the queue to work independent of threads removing elements from the
queue.