Remove usage of Python in MySQL Router unit tests
We remove the need of Python in unit tests. The tests are now
writting in C++ using GTest/GMock.

We remove also the dependency on Python from the RPM and
Debian packages.

Some tests are only able to run when they are started from
with in the Git repository.

(cherry picked from commit 5f26dce)
(cherry picked from commit 1c71f6d)