BUG#22546709: Catch and propagate errors in plugin functions
Before this patch, any exceptions thrown in the plugin would cause the
application to unconditionally terminate. After this patch, exceptions are
caught and propagated to the harness where they can be caught and handled.

The patch accomplishes this by switching from using `std::thread` and instead
dispatch the plugin threads using `std::async`. If an exception is thrown from
one of the plugins, it it caught and the remaining plugins are gracefully
stopped before the exception is re-thrown.

This patch adds a `stop` callback to the plugin interface and step the
interface version accordingly. If a plugin does not implement the stop method,
it is assumed that any threads started can be aborted forcefully when the
application exits or that can be left running during a restart.