BUG22020088 Fix possible issue blocking Router host in MySQL Server
It is possible to block the host from which Router connects with
the MySQL Server. Handshakes which are not responded in a correctly
or timely manner, are logged as an error for each host. Since all
connections through Router are coming from the same host, the host
is blocked more rappitly.

The solution is to catch unfinished handshakes in Router and tell
the MySQL server that all is OK. We do this by checking the
packet number sequence, the length of the handshake response
sent by the client, and the time it takes to send the response.
When the client failed handshaking correctly, Router will send
a fake reply to the MySQL Server, which will make sure it does
not increment the connect errors for the host Router runs on.

Maximum connect or handshake errors can be configured using the
`max_connect_errors` option for each Connection Routing:

  [routing]
  max_connect_errors = 5

Default is set to 100, like the MySQL Server, but unlike the MySQL
Server, the maximum is limited to 2^32 (unsigned int).

It is also possible to set the timeout until Router disconnect and
registers a connect error. The option is called `client_connect_timeout`
which is similar to the MySQL Server option `connect_timeout`:

  [routing]
  client_connect_timeout = 4

Default is set to 9 seconds which is the default of MySQL Server 5.7
minus 1. The minimum is 2, maximum 31536000.

Note that each routing has its own list of blocked host.

Clients which are blocked receive the MySQL Server error 1129
with a slighly different error message:

1129: Too many connection errors from fail.example.com

The Router logs will contain extra information on which clients
are blocked:

INFO    [...] 1 authentication errors for fail.example.com (max 100)
WARNING [...] blocking client host fail.example.com

For now, a restart of Router is necessary to unblock hosts.
(cherry picked from commit 456450d)