Fix legal tests using master branch and gcoverage.sh
We make sure we use the master branch to check the copyright years.
This is to prevent mismatch when commits are cherry picked or pushed
to release branches.

We also fix the gcoverage.sh script so it can run without tests
enabled or run.
(cherry picked from commit a967c65)