BUG22689030 Do not handle MySQL errors as connection errors
Host running MySQL clients would get blocked when receiving MySQL
Server error 1040 (Too many connections) when the MySQL Server is
configured to not allow more connections than what the Connection
Routing plugin allows.

We fix this by checking errors coming before the MySQL Server handshake.
When an error is received, we are not considering this a connection
error and pass the error to the client.

We also increase FD_SETSIZE to 4096 to allow testing and, allow usage
of more file descriptors.