BUG23183566 Reset blocking after connection was established
An issue could occur when uploading data to a MySQL server in chunks,
for example, when inserting data into a BLOB field. The Connection
Routing plugin would terminate the connection with a 'write error'.

We fix the issue by correctly setting the socket to blocking after
it was set to non-blocking when requesting a connection from a
MySQL server.
(cherry picked from commit 16a5082)