Fix Routing plugin startup if socket file already exists
Up until now, if Routing plugin was configured to run using a UNIX
socket and a socket file was leftover from previous run, it would
try to detect if the socket file can be reused.  It did that by
testing if it can bind to it.  Unfortunately, the check didn't work
properly.

To fix this, we changed the assumptions.  Now, the socket file is
expected to be cleaned up when the Router terminates.  Upon a fresh
start it should not exist, and if does, the Router will exit with
error.

We also applied minor code fixes in a few places (they do not affect
program logic).