BUG22705168 Fix incrementing active routes before spawning thread
The counter for active routes in the Routing plugin was incremented
after the client and server sockets were checked and valid. This meant
that when getting the server socket could take some time, the next
client connection would have been accepted although the max number of
connections was reached.

We make sure that the counter is incremented before we spawn the thread
in the Routing plugin.