Fix tests using exotic out-of-source building
Some tests are using the Git source repository to fetch
information such as the year of the last commit to check the
copyright statements. This, however, did not work well when the
out-of-source build directory was not inside the source.

We fix the dection of the source directory to have a better error
messages, and skip tests when Git repository is not available.
For runs internally, we support PB2 environment variables so
automatic builds can run tests.

We also fix the missing comma in the copyright statement when
executing `mysqlrouter --help`.