Adding cpplint checking and pre-commit and update hooks for Git.
This patch add target to run cpplint and check that the coding style is
followed. It also include changes to files necessary to ensure that the check
pass. When searching for 'cpplint' it searches for both 'cpplint' and
'cpplint.py'.

This patch also adds an internal directory that contain a client-side
pre-commit hook as well as a server-side update hook.

Prior to accepting a patch, both hooks will check:

- That there are no trailing whitespaces on any lines.

- That no file names contain non-ascii characters.

- That all C++ source and header files pass the cpplint run.