BUG 24296118 - WRONG IPS ARE LOGGED WHEN USER TRIES TO CONNECT TO A UNREACHABLE FABRIC GROUP
When the router can't connect to MySQL server using rules from the
configuration it logs and also sends the error message to the client:
'Can't connect to MySQL server on <IP address>' giving the IP address of
the router's port on which the client attempted the connection.
This can be perceived as misleading as when for example the router runs
on the same machine as the client issuing:
./bin/mysql -uroot -P 7001 --protocol=TCP
can yield error:  'Can't connect to MySQL server on 127.0.0.1'
even though there is no server (nor the routing) on the local machine.

I've removed the IP address from this message, both in the client error
and in the router's log.