Fix read-write mode failover
This patch fixes the manner in which Routing plugin selects next server
when current server goes down, while running read-write mode.

By design, when running in this mode, failover should progress down
the list of destinations until last of the servers is no longer
available. After that point, it should no longer try any more servers
and refuse to connect, regardless of whether the previously-failed
servers went back online.

Before this patch, the bahavior was different: after traversing the
entire list, if the client tried reconnecting twice, on the 2nd attempt
the failover "wrapped around" back to the first server and traversed the
list again. The client was thus able to reconnect back to any resurrected
servers.

For example, if the list of destinations are servers {A,B,C} (in that
order), the expected failover traversal would be:

  A -> B -> C -> refuse to connect from now on

Before this patch, behavior actually was:

  A -> B -> C -> conn error -> A -> B -> C -> conn error -> A -> B .. etc