Fix for BUG24316097 - SEG FAULT WHEN FABRIC DEST ADDR IS HAVING INCOMPLETE ALLOW_PRIMARY_READS OPTION
Using invalid(incomplete) query in the destination URI lead to array out of
bounds access attempt. Although vector's at() method was used the thrown
exception was never caught leading to abort.

I've added verification before accessing the vector and replaced the exception
with the one that is already caught and handled by the calling code.