Fix for BUG23857183 UNREACHABLE DESTINATION CONNECTION REQUESTS ARE NOT HANDLED PROPERLY IN WINDOWS
When the router tries to connect to the server it uses non blocking connect
and then the select with a timeout.
What's expected is that when the connection fails the select will return
immediately with read descriptors marked as ready.
That is how it works on Linux but on Windows it takes the whole tiemout for
the select to return.

The solution is to add the third descriptors set to the select (exceptfds).
That causes the select to return with exceptfds marked as ready and lets
to avoid waiting the whole timeout period.