BUG24527866 Adding both library path and plugin path to rpath
In order to be able to load both plugins (from other plugins) and libraries
(from plugins and executables) it is necessary to set the rpath for the
executable to contain both the library path and the plugin library path, but
currently only a single entry is present. For installation layouts not
requiring relative paths, this is not a problem since the full path to the
libraries will be used, but for layouts needing relative paths, this will break
and report an error that the shared library cannot be loaded.

This patch remove the `HARNESS_PLUGIN_RPATH` and set `CMAKE_INSTALL_RPATH` to
contain both directories for all layouts requiring relative rpaths. To provide
correct support on Mac OSX, it also set the `CMAKE_INSTALL_NAME_DIR` to
`@rpath` for installation layouts requiring relative paths.

The patch also contain a correction of the output directory for the harness
library on non-Windows builds that were incorrectly set to `<prefix>/bin`
instead of `<prefix>/lib` as part of adding Windows support.