Bug24578638 Fix for invalid code writing data to the receiver socket
The code that was writing the packets read from the sender (client or server)
to the receiver was invalid. If the single write() call did not write all
requested data, the whole buffer was written in another attempt.

This patch fixes this problem, only remaining part of the buffer gets written.
This patch also enables more unit tests for the routing code moving some of the
most used socket related system functions to the separate class that makes them
easily mockable.