BUG24695503 Removing use of INSTALL_LIBDIR in packaging scripts
Bug was a result of setting `INSTALL_LIBDIR` to an absolute path during "cmake"
when executing packaging scripts. The variable `INSTALL_LIBDIR` is set to a
path relative the install prefix. Since the RPATH is computed as
`${CMAKE_INSTALL_PREFIX}/${INSTALL_LIBDIR}/mysqlrouter` this will result in an
incorrect RPATH for the executable and libraries.